+++ src/DependencyInjection/Compiler/ListCompilerPass.php
@@ -1,0 +1,43 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2019 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\ListBundle\DependencyInjection\Compiler;
+
+
+use HeimrichHannot\ListBundle\Registry\ListConfigElementRegistry;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+class ListCompilerPass implements CompilerPassInterface
+{
+
+    /**
+     * You can modify the container here before it is dumped to PHP code.
+     */
+    public function process(ContainerBuilder $container)
+    {
+        // always first check if the primary service is defined
+        if (!$container->has(ListConfigElementRegistry::class)) {
+            return;
+        }
+
+        $definition = $container->findDefinition(ListConfigElementRegistry::class);
+
+        // find all service IDs with the app.mail_transport tag
+        $taggedServices = $container->findTaggedServiceIds('huh.list.config_element_type');
+
+        foreach ($taggedServices as $id => $tags) {
+            // add the transport service to the TransportChain service
+            $definition->addMethodCall('addListConfigElementType', [new Reference($id)]);
+        }
+    }
+}