+++ src/DependencyInjection/ListExtension.php
@@ -34,5 +34,6 @@ class ListExtension extends Extension
 
         $loader->load('services.yml');
         $loader->load('listener.yml');
+        $loader->load('datacontainers.yml');
     }
 }