+++ src/Model/ListConfigElementModel.php
@@ -19,6 +19,7 @@ use Contao\Model;
  * @property int    $dateAdded
  * @property string $title
  * @property string $type
+ * @property string $templateVariable
  * @property string $imageSelectorField
  * @property string $imgSize
  * @property string $placeholderImage